requirements.txt
asgiref==3.10.0
azure-core==1.36.0
azure-storage-blob==12.27.1
bleach==6.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.2.7
django-markdownify==0.9.5
django-storages==1.14.6
gunicorn==23.0.0
idna==3.11
isodate==0.7.2
Markdown==3.10
packaging==25.0
pillow==12.0.0
pycparser==2.23
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.3
tinycss2==1.4.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
whitenoise==6.11.0
